+++ frontend/src/constants/PU/puMessage.ts
@@ -1,5 +1,5 @@
 export const PUMessage = {
-  KAKAO_BTN_NAME: '카카오로 시작하기',
+  KAKAO_BTN_NAME: '카카오 로그인',
   KAKAO_BTN_URL_NOT_EXIST: '카카오 로그인 URL이 존재하지 않습니다.',
   KAKAO_BTN_LOGIN_FAIL: '카카오 로그인 중 오류가 발생했습니다.',
 }
