+++ frontend/src/pages/PU/PU-002/index.tsx
@@ -1,11 +1,8 @@
-import { LoginForm } from './LoginForm'
 import { Wrapper } from '../../../components/layout/Wrapper'
-import { KaKaoButton } from '../components/ui/Button'
 import { LoginMessage } from '@/constants/PU/loginMessage'
+import { KaKaoButton } from '../components/ui/KakaoButton'
 
 export default function LoginPage() {
-  const lineClass = 'border-t border-ccLine my-5'
-
   return (
     <div className="wrap">
       <Wrapper title={LoginMessage.TITLE} className="px-8">
